fix(graphiql): rename files to kebab case
