+++ packages/grafbase-graphiql/src/hooks/use-merge-query.ts
