+++ packages/grafbase-graphiql/src/hooks/use-prettify-editors.ts
