+++ packages/grafbase-graphiql/src/utils/rename-tabs.ts
