+++ packages/grafbase-graphiql/src/utils/validate-query.ts
